+ Index.html first implementation + .gitignore file + init version code refactoring
