+++ model/newsTypeEnum.js
