+++ package.json
@@ -8,12 +8,18 @@
   },
   "keywords": [],
   "author": "Krzysztof Juchnowicz",
+  "contributors": [],
   "license": "MIT",
   "dependencies": {
     "body-parser": "^1.18.2",
+    "bootstrap": "^3.3.7",
+    "bootstrap-select": "^1.12.4",
+    "bootstrap-table": "^1.11.2",
     "cheerio": "latest",
     "cookie-parser": "^1.4.3",
     "express": "^4.16.2",
+    "jquery": "^3.2.1",
+    "jsdom": "^11.5.1",
     "mongodb": "^3.0.1",
     "mongodb-stitch": "^2.3.1",
     "multer": "^1.3.0",
